share code via 'frontend-shared' module
