+++ packages/frontend-shared/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "@packages/frontend-shared",
+  "version": "0.0.0-development",
+  "private": true,
+  "main": "index.js",
+  "browser": "src/index.ts",
+  "scripts": {
+    "build-prod": "tsc",
+    "clean": "rm -f src/*.js src/**/*.js"
+  },
+  "dependencies": {},
+  "devDependencies": {
+    "typescript": "^4.2.3"
+  },
+  "files": [
+    "src/*"
+  ],
+  "types": "src/index.ts"
+}
