+++ packages/launchpad/src/main.ts
@@ -3,7 +3,7 @@ import './main.scss'
 import 'virtual:windi.css'
 import urql from '@urql/vue'
 import App from './App.vue'
-import { makeUrqlClient } from './graphql/urqlClient'
+import { makeUrqlClient } from '@packages/frontend-shared/src/urqlClient'
 import { createI18n } from './locales/i18n'
 
 const app = createApp(App)
